+++ CHANGELOG.md
@@ -20,6 +20,7 @@ Breaking changes:
   example), managing a mouse emulation, or any other ideas you can
   have. As there is a default value for the type parameter, the update
   should be transparent.
+* Layers don't sum anymore, the last pressed layer action set the layer.
 * Rename MeidaCoffee in MediaCoffee to fix typo.
 
 # v0.1.1